+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 - Added `craft\commerce\events\ModifyPurchasablesQueryEvent`.
 - Added `craft\commerce\controllers\OrdersController::EVENT_MODIFY_PURCHASABLES_QUERY`.
 - Guest customers registering during checkout now have their addresses saved to their account.
+- Product conditions can now have a “Variant SKU” rule.
+- Product conditions can now have a “Variant Has Unlimited Stock” rule.
+- Product conditions can now have a “Variant Price” rule.
+- Product conditions can now have a “Variant Stock” rule.
 - Deprecated `craft\commerce\elements\Order::setEmail()`. `Order::setCustomer()` should be used instead.
 - Added `craft\commerce\elements\db\ProductQuery::$shippingCategoryId`.
 - Added `craft\commerce\elements\db\ProductQuery::$taxCategoryId`.